Replace broken AccessTokenManager
https://github.com/DEFRA/water-abstraction-team/issues/74

Whilst digging into an issue with the bill run process we spotted a _few_ (hundreds!) of log messages about the access token being invalid when trying to authenticate. So, we started digging into that. What we realised is that it's, erm, broken!

The previous team chose to take advantage of [Got's hooks](https://github.com/sindresorhus/got/blob/main/documentation/9-hooks.md) to add the authentication header. So, we have this [beforeRequest() hook](https://github.com/sindresorhus/got/blob/main/documentation/9-hooks.md#beforerequest).

```javascript
const beforeRequestHook = async () => {
  if (!accessTokenManager.isTokenValid()) {
    logger.info('Fetching a new Cognito token')
    // Save for further requests
    const accessToken = await accessTokenManager.refreshAccessToken()
    instance.defaults.options.headers.Authorization = `Bearer ${accessToken}`
  }
}
```

Can you spot the problem!? 😁

Should the request return a [401 Unauthorized](https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/401) an [afterReponse() hook](https://github.com/sindresorhus/got/blob/main/documentation/9-hooks.md#afterresponse) steps in. It requests another access token from the [sroc-charging-module-api](https://github.com/DEFRA/sroc-charging-module-api), amends the `Authorization` header and tries again.

```javascript
const afterResponseHook = async (response, retryWithMergedOptions) => {
  // ...

  if (response.statusCode === 401) { // Unauthorized
    // Refresh the access token
    const accessToken = await accessTokenManager.refreshAccessToken()
    const updatedOptions = {
      headers: {
        Authorization: `Bearer ${accessToken}`
      }
    }

    // Save for further requests
    instance.defaults.options = got.mergeOptions(instance.defaults.options, updatedOptions)

    // Make a new retry
    return retryWithMergedOptions(updatedOptions)
  }

  // No changes otherwise
  return response
}
```

So, there are a number of problems, both with the approach and its implementation. Going back to the `beforeRequest()` hook the first issue is it is only setting the `Authorization` header when the access token is invalid. If the `AccessTokenManager` has a valid token, we don't set it!

Next, how it's setting the header differs from the documentation.

```javascript
// Edited slightly to be an actual function
function beforeRequest (options) {
  options.body = JSON.stringify({payload: 'new'});
	options.headers['content-length'] = options.body.length.toString();
};
```

So, we're not updating the options passed in. The other interesting thing about our approach, is it differs with how we did things in `src/lib/connectors/charge-module/lib/got-with-proxy.js`, which by the way, follows the docs. This one we've not confirmed but our belief is that it's not working.

The service is saved though by the `afterResponseHook()`. What this means is the request lifecycle seems to be

- beforeHook()
  - sees the token is invalid
  - *REQUEST* sends request to AWS Cognito for a token
  - fails to add auth header correctly
- *REQUEST* send our request to the Charging Module API
- `401` response
- afterResponseHook()
  - sees the 401 response
  - *REQUEST* sends additional request to AWS Cognito for another token
  - succeeds in adding auth header
- *REQUEST* retry the original request to the Charging Module API
- `2xx` response

What can we say. In the legacy code we _think_ we're making a single request (maybe 2 for the very first one). But instead we're making 4!😮

Next we're running the the service on 2 separate servers in our AWS environments. So, we accept that each instance will need to manage its own token. But the `AccessTokenManager` is instantiated in the scope of a BullMQ worker processing a job. So, we believe each job creates it's own instance rather than there being a singleton or global instance of the `AccessTokenManager` and the current access token.

This means if we are dealing with a bill run with 1000 transaction lines, a job is created for each to add it to the bill run in the Charging Module. If we are right each transaction requires four HTTP requests, plus the delay waiting for the `401` response and the retry. 😱

Is it any wonder this thing is as slow as treacle and wants _all_ the resources!? 😩

So, this change fixes the issues with the `beforeRequestHook()`. We also make use of [Node's `global` object](https://nodejs.org/api/globals.html#global) to store the token to reduce the number of requests being made.