Update (rewrite!) AccessTokenManager
All a caller wants is the token. It shouldn't be asking the token manager if a token is valid and then directing the manager to refresh the token.

A cleaner API would just for the caller to request a token, and the token manager determine if it needs to get a new one or not.

So, we have rewritten the AccessTokenManager on that basis. That also meant a rewrite of the tests, which were focused on how it was interacting with the CHA module anyway, and not on what the expected behaviour should be.

We've also been able to make some other improvements, for example, removing the dependence on Moment for the date comparison and date addition. Handy, as this Stackoverflow post highlights Moment is less performant than vanilla JS.

https://stackoverflow.com/a/35153097/6117745