Fix tests in github CI
No idea what is so different between the previous implementation and this, but now build is failing with

```
Invalid URL
      at new NodeError (node:internal/errors:387:5)
      at URL.onParseError (node:internal/url:565:9)
      at new URL (node:internal/url:641:5)
      at AccessTokenManager._refresh (/home/runner/work/water-abstraction-service/water-abstraction-service/src/lib/connectors/charge-module/lib/AccessTokenManager.js:45:192)
      at AccessTokenManager.token (/home/runner/work/water-abstraction-service/water-abstraction-service/src/lib/connectors/charge-module/lib/AccessTokenManager.js:16:177)
```

Should add, its not the switch from `url-join` to `URL`. This error was also happening with it. It's this error that made us see we were reliant on a dependency.