Make Got http request retries more robust (#2000)
https://github.com/DEFRA/water-abstraction-team/issues/74

One of the other things we've spotted with the legacy billing process is it's using both [Got](https://github.com/sindresorhus/got) and [BullMQ](https://docs.bullmq.io/) to retry failed requests.

In [Fix broken AccessTokenManager](https://github.com/DEFRA/water-abstraction-service/pull/1993) we've hopefully nullified the repeated `401` errors we see in the logs.

But sometimes a request will timeout, for example when requesting a token from the AWS Cognito service for the first time (it is like you kick it awake on the first attempt each day 🤷🫤). With the current setup, it's not **Got** that retries these, but **BullMQ** (if configured in the job).

Using our experience on [sroc-charging-module-api](https://github.com/DEFRA/sroc-charging-module-api) we believe there are changes we can make to **Got** so where we genuinely need to re-try, it can be done there. We can then remove the retry mechanism from the **BullMQ** jobs that don't need it. But we'll make those changes in a separate PR.

---

We're not sure why general options that are not relevant to dealing with a proxy are in `got-with-proxy.js`. So, it seemed to make more sense to move them into our general got-cm module.