Tidy up onFailedHandler
There was logic in the `onFailedHandler()` for dealing with whether it was the last attempt that the job failed. As we are no longer retrying the job this becomes invalid. But we also spotted that this function is using the logger directly, instead of the helpers in `src/modules/billing/jobs/lib/batch-job.js`.

So, we also took the opportunity to tidy things up, and make use of an existing helper for both logging an error and setting the error status code on a batch.