Revert and move Got config (#2002)
https://github.com/DEFRA/water-abstraction-team/issues/74

In the recent change [Make Got http request retries more robust](https://github.com/DEFRA/water-abstraction-service/pull/2000) we remarked that config we would expect to see in `src/lib/connectors/charge-module/lib/got-cm.js` was in `got-with-proxy.js`. So, we moved it!

We very quickly realised why it was there when everything broke! 🤦

What we _should_ have spotted is that `src/lib/connectors/charge-module/lib/AccessTokenManager.js` uses `got-with-proxy.js` because of the hooks built into `got-cm.js`. These automatically call AccessTokenManager, which would lead to a right mess if we used it.

So, when we moved the config we broke AccessTokenManager. This means what we _should_ have done is move our general config down into `got-with-proxy.js`. Then both calls to the Charging Module API and the AWS Cognito instance will benefit from our more robust settings.