Stop hammering DB for batch state during bill run
https://github.com/DEFRA/water-abstraction-team/issues/74

The legacy code, because of its use of [BullMQ](https://docs.bullmq.io/) is left with a bit of a problem to crack. Each step of the bill run process is represented as a 'queue' in **BullMQ**. To kick things of a job is added to the first queue (step) in the process. Typically, when one job completes it handles adding the next job to the next appropriate queue in the process.

But the point of BullMQ was for steps like creating transactions in the charge module; there could be thousands. So, you add all of them as individual jobs to a queue, and you have multiple workers process them asynchronously.

What happens if the next step should only happen once? For example, after adding your thousand transactions you need to tell the [Charging Module API](https://github.com/DEFRA/sroc-charging-module-api) to 'generate' the bill run. But you only need to send that request once.

1000 create-charge jobs need to somehow queue up a single refresh-totals job in their [onComplete() event handler](https://docs.bullmq.io/guide/returning-job-data).

The legacy solution is to query the DB each time and get a count of the different states of the transactions linked to the bill run. The more transactions you have, the more times you fire that query and the more work it has to do.

What if the last create-charge job knew it was the last one added to the queue? By default jobs are processed on a [First-In, First-Out](https://docs.bullmq.io/guide/jobs/fifo) basis. So, the last job would know it was the last to be processed. It could then queue up the refresh-totals job and we'd avoid all those hits on the DB.

This change implements this new approach to dealing with when to queue the refresh-totals job.