Update createCharge createMessage() with new flag
We now have a new flag which we'll be using to denote whether this is the last create-charge job that got added. We only need to do our clean up and queue the next job in the last one. This will be the thing we use to know when to do that.