Update create-charge to stop querying DB
We still have the work to do of setting the `lastOfUs` flag. But now we have it we can update the create-charge job to look out for it rather than keep hitting the DB to ask whether all the transactions are done.

This also applies to the clean up job that was hidden away in the concept of checking whether creating charges are done. We can now just call it directly in the onComplete.

We've also tweaked the order of things a bit. We now tell the Charging Module to start generating. We then clean up the batch before queuing the refresh-totals job. For small bill runs it is much more likely the CM will have finished generating the bill run by the time the refresh-totals job kicks off, which should reduce the instances of 'CM bill run summary not ready for batch' we seeing being logged.