Stop hammering DB for batch state during bill run (#2003)
https://github.com/DEFRA/water-abstraction-team/issues/74

The legacy code, because of its use of [BullMQ](https://docs.bullmq.io/) is left with a bit of a problem to crack. Each step of the bill run process is represented as a 'queue' in **BullMQ**. To kick things off a job is added to the first queue (step) in the process. Typically, when one job completes it handles adding the next job to the next appropriate queue in the process.

But the point of BullMQ was for steps like creating transactions in the Charging Module; there could be thousands. So, you add all of them as individual jobs to a queue, and you have multiple workers process them asynchronously.

What happens if the next step should only happen once? For example, after adding your thousand transactions you need to tell the [Charging Module API](https://github.com/DEFRA/sroc-charging-module-api) to 'generate' the bill run. But you only need to send that request once.

1000 create-charge jobs need to somehow queue up a single refresh-totals job in their [onComplete() event handler](https://docs.bullmq.io/guide/returning-job-data).

The legacy solution is to query the DB each time and get a count of the different states of the transactions linked to the bill run. The more transactions you have, the more times you fire that query and the more work it has to do.

What if the last create-charge job knew it was the last one added to the queue? By default, jobs are processed on a [First-In, First-Out](https://docs.bullmq.io/guide/jobs/fifo) basis. So, the last job would know it was the last to be processed. It could then queue up the refresh-totals job and we'd avoid all those hits on the DB.

This change implements this new approach to dealing with when to queue the refresh-totals job.

---

Other notes

- Add logInfo batch job helper

We wanted to log some stuff, just for debugging purposes that included the job ID. We assumed `src/modules/billing/jobs/lib/batch-job.js` would have something but found it didn't.

Then we realised. The previous team built `batch-job.js` with all these helpers that allow you to track log messages across jobs. But unless it's a specific handler, onComplete or error message they revert back to the standard logger. This means you can't link the info messages to the job that spat them out! 🤦😩

So, we added one! 😁

We then applied it in the 2 jobs we were touching.

- Clean up the create-charge tests

The tests don't depend on creating dynamic GUIDs so what's the point of bringing in another dependency to generate them? None. So, swapped them out for fixed GUIDs.

We also found a lot of duplication of data, some that weren't being used, and some invalid test titles (a clear case of copy and paste)

- Clean of work to do in create-charge

Because we know when we are handling the last create-charge job on the queue, we do not have to keep querying the DB to find out if we are 'ready'. The old logic also triggered cleaning up the batch when we reached a 'ready' state.

Obviously, the aim of this change was to remove the hit on the DB. But we can also tidy things up and move the clean-up call to our onComplete, triggering when we have reached the `lastOfUs`.

We now have a new flag which we'll be using to denote whether this is the last create-charge job that got added. We only need to do our clean-up and queue the next job in the last one. This will be the thing we use to know when to do that.

We've also tweaked the order of things a bit. Now we tell the Charging Module to start generating. We then clean up the batch before queuing the refresh-totals job. For small bill runs it is much more likely the CM will have finished generating the bill run by the time the refresh-totals job kicks off, which should reduce the instances of 'CM bill run summary not ready for batch' we see being logged.