Fix no contacts in external login
https://github.com/DEFRA/water-abstraction-team/issues/25

A recent change as part of our efforts to remove the code base's dependency on Lodash has introduced an error. The Lodash `get()` method doesn't care if the object you are getting a property from is undefined. It will just return whatever default value you provide.

Our issue is that the existing code nor the tests gave any clue this might be the case. So, though we test for the property `name` we were assuming `doc.metadata.contacts[0]` would always be defined.

Turns out this is not the case. So, this change fixes the issue and prevents an error from being thrown.

For reference the error we saw was `Uncaught TypeError: Cannot read property '0' of undefined`.