Remove 'isObject', 'first', 'isUndefined', and 'find' lodash function (#1998)
https://github.com/DEFRA/water-abstraction-team/issues/25

Lodash is used a lot in our codebase. Although we are not totally removing this module, some of the functions in the code using lodash are unnecessary. Here we are removing the unnecessary ones are replacing it with vanilla js.
This PR is focusing on the 'isObject', 'first', 'isUndefined', and 'find' functions