Add new batch refresh totals endpoint (#2036)
https://eaflood.atlassian.net/browse/WATER-3906
https://eaflood.atlassian.net/browse/WATER-3912

> This is part of the [SROC supplementary bill run work](https://eaflood.atlassian.net/browse/WATER-3803)

In the supplementary bill run process the [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system) is taking things as far as having the batch, invoice, invoice licence and transaction records created in the DB and the Charging Module API.

It will then hand over control to the legacy service to extract ‘post-generate’ info from the charging module and update our data with it, and all other actions from that point on.

So, we need a way of passing control to the legacy service. Currently, in the `src/modules/billing/jobs/create-charge.js` job, once the transactions have been sent to the charging module and the bill run generated, it adds the `src/modules/billing/jobs/refresh-totals.js` job to a queue.

We need to add a new endpoint that allows us when called, to add that **refresh-totals** job for our newly created SROC supplementary batch.