Refactor removal of lodash code
https://github.com/DEFRA/water-abstraction-team/issues/25

Lodash has been used a lot in our code base. We have been working hard to remove all of the lodash code and replace it with vaniilla javascript instead. In doing this there has been a small number of lodash functions that we have needed to keep ( certaintly not all the ones that had been used). These had been put into there own modules to be called around the code base. After the 3rd module was created we decided it would be tidyer to keep all the lodash modules in its own file and export them there. This PR does that.