Add SROC include in supplementary billing flag
https://eaflood.atlassian.net/browse/WATER-3948

We have an issue that the current `include_in_supplementary_billing` was added at a time there was only 1 charge scheme. A licence gets flagged irrespective of whether the change relates to PRESROC or SROC.

Where that has impacted us is when sending a billing batch. When it gets sent it clears the flag for _all_ licences included. But if we cancel one, for example the PRESROC bill run, and send the SROC one we loose which licences still need to be processed on the PRESROC one.

Our solution, is to add a new `include_in_sroc_supplementary_billing` flag to the `licences` table. This change covers

- a migration to add it to the table
- updating the logic which sets the include in supp. billing flags when a charge version is approved to update only the relevant flag