Revert Puppeteer to 19.8.0 (#2094)
https://github.com/DEFRA/water-abstraction-team/issues/80

Bumping puppeteer has caused it to try to download a later version of Chromium. This is fine locally and in CI but it is failing in our AWS environments. Most likely there is a firewall or something blocking it but we need to ship a release to production so we are reverting it to the last known working version and will look into it properly when we have more time.