Supplementary Billing - Too many previous years
https://eaflood.atlassian.net/browse/WATER-3944

The initial fix for this bug created a new issue where you could re-create another “old scheme” bill run when there was one already sitting in `READY` status for that region.

This was caused by moving the code which calculates the ALCS `toFinancialYearEnding` below where it checks if the batch already exists. This results in `getExistingOrDuplicateSentBatch` function using the wrong `toFinancialYearEnding` value to check for an existing batch.