Fix bug in acceptance test fixture loader (#2124)
We recently made a change to the fixture loader that is used to create test data from YAML files. The issue was simply we thought we were iterating an array whereas we are iterating the properties of an Object.

So, a quick change from `of` to `in` and all sorted. But what made it a pain was the poor choice of names for things, for example, using `config` in relation to two completely different entities.

So, we have also made a stab at trying to make things clearer which is why the change is larger than just the fix. It's not perfect, but at least we're not confusing `data` with `data` and `config` with `config` anymore 😩🤦😁