Supported source charge is not correct (#2142)
https://eaflood.atlassian.net/browse/WATER-4001

When viewing the transactions for a licence. If the transactions include Additional charges it has been noticed that whilst the supported source name is correct the charge is not correct.

After some digging, it has been found that the old code is getting this value from the "rules engine" data that is returned by the charging module. This data is never intended for use for anything other than debugging, and for the `supportedSourceCharge` it appears to return its gross value, not the net.

The correct value is returned as part of the response from the charging module. So this PR is going to correct the location in the response that this data item is being pulled from.