Add migration to delete empty & errored bill runs (#2152)
We made a change to the UI that finally allows users to [delete empty and errored bill runs](https://github.com/DEFRA/water-abstraction-ui/pull/2349).

So any new ones that get added can be dealt with. Unfortunately, we have loads of existing ones that unnecessarily drag down the performance of the bill runs page and make it harder to find meaningful bill runs.

So, this change adds a migration that will do a 'clean-up' of these empty and errored billing batches when deployed. It will be on the users to delete any new ones created.