Add migration to unflag non-PRESROC licences
https://eaflood.atlassian.net/browse/WATER-3995

We added a migration recently to [unflag non-SROC licences](https://github.com/DEFRA/water-abstraction-service/pull/2150). At the same time we learnt that the [water-import-service](https://github.com/DEFRA/water-abstraction-import) was setting `include_in_supplementary_billing` for any licence where a change in the revoked, lapsed or expired date was noticed. It wasn't checking if the licence has only SROC charge versions, or if it had none.

This causes the same problem our previous migration was fixing; licences get flagged for PRESROC supplementary billing but there is no way to clear it because they'll never be processed when billing runs.

We've [updated the water-abstraction-import](https://github.com/DEFRA/water-abstraction-import/pull/631) to be more intelligent in how it flags licences. This migration will remove the flag from those licences that should never have been flagged in the first place.