Empty supplementary bill run showing a credit bill
https://eaflood.atlassian.net/browse/WATER-4000

When a bill run is created that contains invoices, and those invoices are then removed from the bill run. Upon removal of the final invoice from the bill run it is correctly set to "empty" but still shows 1 bill in the bill run along with the value of the last invoice.

When invoices are removed it checks to see if there are any left in the batch. If the number of invoices was 0 it was setting the batch status to “empty” but that was it, it didn’t update any of the values. So the billing batch record retained the values of the last invoice it was linked to since these weren’t cleared out.

This PR will fix this by clearing the remaining invoice data at the same time as setting the billing batch `status` to "empty".