Tell Got to retry PATCH requests (#2204)
https://eaflood.atlassian.net/browse/WATER-4055

We spotted this when an annual bill run after being 'sent' appeared to become stuck. The service failed with a timeout whilst firing the `/approve` and `/send` requests to the [Charging Module API](https://github.com/DEFRA/sroc-charging-module-api). The fact the CHA responded with `204` to both requests immediately is a thing to look into another day!

But WRLS thought it got a timeout and then blew up, leaving the bill run in a state of sending but the bill run in the Charging Module flagged as 'billed' and on its merry way to be sent out.

Things _might_ have been ok if it could have retried the request. If nothing else it brings the config in line with what we do in [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system).

Hence we're making this change.